fix testpmd to run without Xen environment
Note: upstream discussion is slow, but we need a fix now. We will drop
this and change to the upstream solution when it is available.

With LIBRTE_PMD_XENVIRT enabled testpmd is built in a way to ONLY work
in XEN environments.
It will surface as:
   PMD: gntalloc: ioctl error
   EAL: Error - exiting with code: 1
     Cause: Creation of mbuf pool for socket 0 failed

There could be a complex solution checking if this is a Xen env via
something like:
   #include <xenstore.h>
   struct xs_handle *xsh = NULL;
   xsh = xs_open(XS_OPEN_READONLY);
   check for xsh being NULL ...

But I see no reason to go so complex and instead the patch goes the
simple and more readable way of just falling back to the normal
allocation if the xen based allocation failed.