avoid failures due to missing modules
This could have happened e.g. in cloud image environments where extra modules
are not default installed.
This makes it fail gracefully, add a comment in the config file the user
is touching and suggests the right meta package to get the drivers.
Since it is working without e.g. via virtio-pmd it would not qualify for
a recommend.