Fix issues with some autopkgtest environments
The Hugepage tests already worked in environements with and without 1G
hugepage support. They were failing in environments with support for 1G,
but not enough memory.
Therefore this patch skips the verification of the 1G alloc (but keeps
the config to do it, as we want to see it fail gracefully).
It also reduced the ammount needed for the 2M hugepages test.
And finally it fixes some vairable references related to the hugepage
tests.