fix handling of multiple huge page sizes
If multiple huge pages sizes were requested it could happen that some of
them had no valid hugetlbfs mountpoint (LP: #1557532).
To solve that dpdk-init had to learn to understand what the default huge
page size is and to take care of mounting this and/or any other size
depending on the requested hugepages in dpdk.conf.