lpm/lpm6: fix use after free and missing frees
In certain autotests lpm->max_rules turned out to be non initialized.
Also lpm6 autotests failed with the default alloc of 512M Memory.
While >=2500M was a workaround it became clear while debugging that
it had a leak.

That eventually led to the identification of various use after free
and leaks
in the lpm/lpm6 code.

Along that is also makes the RTE_LOG messages of the failed
allocation unique.